requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg[binary]
python-dotenv
requests
